+++ zenpredict-bake/tests/bake_roundtrip.rs
@@ -1,6 +1,3 @@
-use zenpredict::*;
-use zenpredict_bake::*;
-
 mod bake_roundtrip {
     use zenpredict::MetadataType;
     use zenpredict::*;